remove tests for query param bug
The tests don't actually catch the error. The Ember test server
does not fully replicate the behavior of triggering partial
and full transitions for routes. The page transition is being
triggered by the Search Mixin when resetPagination is called on
change. We can't test the route in isolation because we're using
the Search mixin and that's where the behavior causing namespaces
to disappear is coming from.